+++ bundles/org.eclipse.swt/Eclipse SWT/cocoa/org/eclipse/swt/widgets/Slider.java
@@ -265,7 +265,7 @@ public int getPageIncrement () {
 public int getSelection () {
 	checkWidget();
 	NSScroller widget = (NSScroller)view;
-	float value = widget.floatValue();
+	double value = widget.doubleValue();
     return (int)((maximum - thumb - minimum) * value + minimum);
 }
